マウスで回転
git-svn-id: file:///home/gecchi/repos@3690 9df8dbd1-7660-4e89-b1e7-fa117e7ea0c7

+++ GgafDx9Core/include/jp/ggaf/dx9core/util/GgafDx9Quaternion.h
@@ -11,10 +11,10 @@ namespace GgafDx9Core {
  */
 class GgafDx9Quaternion : public GgafCore::GgafObject {
 public:
-    float   _t; //‹••”
-    float   _x;
-    float   _y;
-    float   _z;
+    double   _t; //‹••”
+    double   _x;
+    double   _y;
+    double   _z;
 
     GgafDx9Quaternion() : GgafObject() {
         _t = 0;
@@ -23,9 +23,9 @@ public:
         _z = 0;
     }
 
-//    GgafDx9Quaternion(float prm_t, float prm_x, float prm_y, float prm_z);
+//    GgafDx9Quaternion(double prm_t, double prm_x, double prm_y, double prm_z);
 
-    GgafDx9Quaternion(float prm_t, float prm_x, float prm_y, float prm_z) : GgafObject(),
+    GgafDx9Quaternion(double prm_t, double prm_x, double prm_y, double prm_z) : GgafObject(),
         _t(prm_t),
         _x(prm_x),
         _y(prm_y),
@@ -33,17 +33,17 @@ public:
     }
 
 
-    inline void set(float prm_t, float prm_x, float prm_y, float prm_z) {
+    inline void set(double prm_t, double prm_x, double prm_y, double prm_z) {
         _t = prm_t;
         _x = prm_x;
         _y = prm_y;
         _z = prm_z;
     }
-    inline void mul(float t2, float x2, float y2, float z2) {
-        float t1 = _t;
-        float x1 = _x;
-        float y1 = _y;
-        float z1 = _z;
+    inline void mul(double t2, double x2, double y2, double z2) {
+        double t1 = _t;
+        double x1 = _x;
+        double y1 = _y;
+        double z1 = _z;
         _t = t1*t2 - x1*x2 - y1*y2 - z1*z2;
         _x = t1*x2 + t2*x1 + y1*z2 - z1*y2;
         _y = t1*y2 + t2*y1 + z1*x2 - x1*z2;
